+++ app/Http/Controllers/ConditionController.php
@@ -2,7 +2,7 @@
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use App\Officer;
+use App\Corpse;
 use App\Commendation ;
 use App\Catcommendation;
 use Validator;
@@ -145,7 +145,18 @@ class ConditionController  extends  Controller
 
             return redirect(route('conditions.index'));
         }
-
+////////////////////
+        $chech_if_Id_InUse=null;
+        $corpses =Corpse::where('condition_id',$id)->get();
+        foreach ($corpses as $corpse) {
+           $chech_if_Id_InUse= $corpse;
+        }
+        
+        if (!empty($chech_if_Id_InUse)) {
+            Session::flash('error','Entity integrity constraints Enforces, Cannot be deleted !');
+            return redirect(route('conditions.index'));
+        }  
+/////////////////////////
        $condition->delete($id);
 
        Session::flash('success','Condition  deleted successfully.');
